Refactor: Moving evolve logic out of the stepper (#639)
* Refactor: Moving evolve logic out of the stepper
