+++ pokemongo_bot/cell_workers/__init__.py
@@ -4,3 +4,4 @@ from pokemon_catch_worker import PokemonCatchWorker
 from seen_fort_worker import SeenFortWorker
 from move_to_fort_worker import MoveToFortWorker
 from initial_transfer_worker import InitialTransferWorker
+from evolve_all_worker import EvolveAllWorker
